Fix completed for jsdom getContext() issue
