+++ .gitignore
@@ -17,3 +17,6 @@ npm-debug.log
 .*.sw[op]
 .commits.tmp
 gen/
+.nyc_output/
+coverage/
+package-lock.json
