+++ package.json
@@ -65,7 +65,6 @@
     "jsdom-global": "^3.0.2",
     "merge-stream": "^1.0.1",
     "mocha": "^3.4.2",
-    "mocha-jsdom": "^1.1.0",
     "nyc": "^11.2.1",
     "rimraf": "^2.6.1",
     "test-console": "^1.0.0",
