+++ test/PointItem.test.js
@@ -1,5 +1,4 @@
 var assert = require('assert');
-//var jsdom = require('mocha-jsdom');
 var jsdom_global = require('jsdom-global');
 var moment = require('../lib/module/moment');
 var PointItem = require("../lib/timeline/component/item/PointItem");
@@ -8,28 +7,14 @@ var TestSupport = require('./TestSupport');
 
 describe('Timeline PointItem', function () {
   
- // jsdom();
   var now = moment();
 
   before(function() {
     this.jsdom_global = jsdom_global();
-//      "<div id='mynetwork'></div>",
-//      { skipWindowCheck: true}
-//    );
-//    this.container = document.getElementById('mynetwork');
   });
 
-
   after(function() {
-//    try {
-      this.jsdom_global();
-//    } catch(e) {
-//      if (e.message() === 'window is undefined') {
-//        console.warning("'" + e.message() + "' happened again");
-//      } else {
-//        throw e;
-//      }
-//    }
+    this.jsdom_global();
   });
 
   it('should initialize with minimal data', function() {
